Moving snapshot testing helper method to their own files
